requirements.txt
# Core
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
pymongo==4.6.3

# Utilities
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.2.0

# Logging & monitoring
gunicorn==21.2.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
faker==24.3.0

# Linting
flake8==7.0.0
flake8-docstrings==1.7.0

# Type checking helpers
typing-extensions==4.10.0
